+++ packages/sdk/src/fluent/index.js
@@ -1,0 +1,1 @@
+export { FluentInstance } from './instance.js';