+++ packages/sdk/src/index.js
@@ -1,5 +1,6 @@
 export * from './client.js';
 export { default } from './client.js';
+export * from './fluent/index.js';
 export * from './drivers/adapter.js';
 export * from './drivers/mediasession.js';
 export * from './drivers/dom.js';